requirements.txt
kafka-python
matplotlib
textblob
Faker
wheel
streamlit
pandas
numpy
scikit-learn
joblib
plotly
xgboost
lightfm
